Fix store catalog
