+++ src/components/Home/Home.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import './Home.css';
 import { useNavigate } from 'react-router-dom';
-import Catalog from '../Catalog/Catalog';
+import Catalog from './Catalog/Catalog';
 
 function Home({ role }) {
   const navigate = useNavigate();
